Add an experiment with config overrides
We want our customers to be able to use *all* Envoy features, but we
won't necessarily be able to build explicit support for them in the
EGW. We need some sort of mechanism that allows customers to make
ad-hoc configuration changes their Envoy instances.

Envoy has the "Runtime" idea with "layers." This commit is an
experiment that sets up three levels of layers that you can play
around with. The first level is the config served by the
xds_operator. That's the baseline. The next level is config key-value
pairs written to a filesystem in the Envoy container. The highest
level is the Envoy admin interface.

Once you create a service and get its pod id, you can exec bash into
that pod to experiment with the filesystem overrides. For example:
 $ kubectl exec -n egw-sample -it pod/envoy-sample-acnodal-59f994dddd-f72qb -- /bin/bash -l
 # cd /etc/envoy/overrides
 # mkdir v1/health_check
 # echo 42 > v1/health_check/min_interval
 # ln -s v1 new
 # mv -Tf new current

If you get the envoy admin ui runtime dump you should see your
changes.

You can also set parameters by the web interface, for example:

 $ http POST
 http://localhost:9901/runtime_modify?health_check.min_interval=60