Add a route for pod CIDR traffic
We need to ensure that inter-pod traffic routes over eth0. We set up a
default route to use the True Ingress net1 interface, so we need a
more specific route to send anything going to a pod CIDR address to
eth0.

We use calicoctl to get the pod network CIDR. The name of the calico
CR is currently hard-coded but if we need it to be variable we'll
probably want to pass it into this image from the controller-manager.

This commit also adds docker-entrypoint.sh to this project. We were
pulling it from our custom Envoy images, but now that we're using
upstream Envoy images it's cleaner to have this project not depend on
that one anymore.